Refactor parts of code
Rename a few constants. Add new constant GEARBOX_RATIO which is used to
calculate the POSITION and VELOCITY_CONVERSION_FACTORs.

Make the code less tightly coupled by passing in the controller and axes
in the TeleopDrive constructor, and passing in the motor controller
addresses in Drivetrain constructor.

Remove the unnecessary call to setInverted(boolean) in Drivetrain
constructor.