Switch to arcadeDriveSimple and fix bugs
Fix bug where the right side of the drivetrain would be inverted, which
makes turning go straight and going straight turning.

Fix bug with TeleopDrive.applyDeadband not calculating properly.

Fix bug with the direction of Drivetrain.arcadeDriveSimple, rename it to
Drivetrain.arcadeDrive and remove the old Drivetrain.arcadeDrive as they
were found to be pretty much equal.

Change the default ramping rate for TeleopDrive to 0.375s.