Add range table loading and fix bugs
Add RangeTable class for a searchable range table that can be loaded
from a CSV file. The range table is loaded in RobotContainer and can be
accessed through Shooter.getRangeTable().

Fix a bug in RobotContainer where the code crashes because the shooter's
PID controller was sent to Shuffleboard without a name.